front/minse plase du template

+++ database/migrations/2025_07_03_133550_create_produits_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('produits', function (Blueprint $table) {
             $table->id();
             $table->string('nom')->unique();
-            $table->decimal('prix', 8, 2)->nullable();
+            $table->integer('prix')->nullable();
             $table->integer('seuil_alerte')->default(0);
             $table->timestamps();
         });
